% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgemmacpp.R
\name{gemma_args}
\alias{gemma_args}
\title{gemma argument details}
\usage{
gemma_args()
}
\value{
Returns a tibble of arguments, their default values, and help text.
}
\description{
Returns a tibble detailing the configuration arguments for gemma class.
}
\examples{
gemma_args()

}
